Refactoring: extract module decoder (29-diamond-event)
